requirements.txt
# OKX现货交易系统依赖

# 核心依赖
ccxt==4.2.85
websockets==12.0
python-dotenv==1.0.0
requests==2.31.0

# 数据处理
pandas>=2.1.3
numpy>=1.26.2
pandas-ta>=0.4.71b0

# 技术分析
ta-lib>=0.4.28

# 异步处理
asyncio==3.4.3
aiohttp==3.9.1

# 配置与日志
pyyaml==6.0.1
structlog==23.2.0

# 数据分析与可视化
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.0
scipy==1.11.4
statsmodels==0.14.1

# Web框架与界面
streamlit==1.32.0
fastapi==0.109.0
uvicorn==0.27.1

# 数据处理与存储
sqlalchemy==2.0.25

# 日期时间处理
python-dateutil==2.8.2
pytz==2023.3.post1

# 监控指标
prometheus-client==0.19.0

# 实用工具
tqdm==4.66.1
rich==13.7.0

# 开发工具
tox==4.11.3
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
mypy==1.7.1
flake8==6.1.0

# 日志记录
loguru==0.7.2